1) Refactored convenience functions to other modules 2) Restructured example files.
