+++ qcodes/sweep/__init__.py
@@ -1,2 +1,3 @@
-from .sweep import sweep, nest, chain, szip, getter, setter
+from .sweep import sweep, nest, chain, szip
+from .getter_setter import  getter, setter
 from .sweep_measurement import SweepMeasurement
